Tear down existing code and rework with a plan
Spent a day or so just knocking out logic in pseudocode.  Put
changes at the bottom of the README for now.  Intending on splitting
that out into a seperate doc later when it's less useful.

As of now, button entry with mouse and keyboard are working and firing
correct events, logging down through the event code and hitting the
right spots for calling the individual functions for the various
buttons.

This basically splits out all the logic I was trying to chain in a
single event handler and lets me have a function 'just do one thing'.

More to come.