Commit to freeze mostly working features
At this point I hate everything about this code.  I have reworked it
and tested and tried to catch edge cases, but it's mostly working now.

That said, I'm considering ripping everything out of the stowArithmetic
function and refactor it down so that it's not storing data in the way
that it is.  The two people that have given me feedback have suggested
using a simple array and putting in input in there, then just running
the operation on the array, but I've tried to consider how to do that
and don't seem to be able to wrap my head around it.

I'll keep working at it and get something I'm at least okay with.  As it
stands, this code is sub-marginal at best.