Refactor and refine js
I ended up starting from fresh...again.  I think this is the third
reimagining of this particular script, and it feels a lot less clunky.

It's still not operating correctly, but it is more /readable/.  I'm
still missing functionality for '=' and 'backspace' especially, but I
also have to deal with decimals sanely as well. I still have a lot of
work left on this, but it's at a point where I'm ready to put it on
GH and start making real commits. For completion, I have moved the old
second iteration of the code to old-calculator.js and have included It
in this push.